+++ src/main/java/net/sf/jabref/JabRefPreferences.java
@@ -130,13 +130,8 @@ public class JabRefPreferences {
     public static final String SIDE_PANE_COMPONENT_NAMES = "sidePaneComponentNames";
     public static final String XMP_PRIVACY_FILTERS = "xmpPrivacyFilters";
     public static final String USE_XMP_PRIVACY_FILTER = "useXmpPrivacyFilter";
-    public static final String SEARCH_AUTO_COMPLETE = "searchAutoComplete";
-    public static final String INCREMENT_S = "incrementS";
-    public static final String SEARCH_ALL = "searchAll";
-    public static final String SEARCH_GEN = "searchGen";
-    public static final String SEARCH_OPT = "searchOpt";
-    public static final String SEARCH_REQ = "searchReq";
-    public static final String CASE_SENSITIVE_SEARCH = "caseSensitiveSearch";
+    public static final String SEARCH_MODE_FILTER = "searchModeFilter";
+    public static final String SEARCH_CASE_SENSITIVE = "caseSensitiveSearch";
     public static final String DEFAULT_AUTO_SORT = "defaultAutoSort";
     public static final String DEFAULT_SHOW_SOURCE = "defaultShowSource";
     public static final String STRINGS_SIZE_Y = "stringsSizeY";
@@ -169,9 +164,9 @@ public class JabRefPreferences {
     public static final String AUTO_COMP_FIRST_LAST = "autoCompFF";
     public static final String AUTO_COMPLETE_FIELDS = "autoCompleteFields";
     public static final String AUTO_COMPLETE = "autoComplete";
-    public static final String HIGH_LIGHT_WORDS = "highLightWords";
-    public static final String REG_EXP_SEARCH = "regExpSearch";
-    public static final String SELECT_S = "selectS";
+    public static final String SEARCH_PANE_POS_Y = "searchPanePosY";
+    public static final String SEARCH_PANE_POS_X = "searchPanePosX";
+    public static final String SEARCH_REG_EXP = "regExpSearch";
     public static final String EDITOR_EMACS_KEYBINDINGS = "editorEMACSkeyBindings";
     public static final String EDITOR_EMACS_KEYBINDINGS_REBIND_CA = "editorEMACSkeyBindingsRebindCA";
     public static final String EDITOR_EMACS_KEYBINDINGS_REBIND_CF = "editorEMACSkeyBindingsRebindCF";
@@ -233,7 +228,6 @@ public class JabRefPreferences {
     public static final String DEFAULT_OWNER = "defaultOwner";
     public static final String GROUPS_VISIBLE_ROWS = "groupsVisibleRows";
     public static final String DEFAULT_ENCODING = "defaultEncoding";
-    public static final String SEARCH_PANEL_VISIBLE = "searchPanelVisible";
     public static final String TOOLBAR_VISIBLE = "toolbarVisible";
     public static final String HIGHLIGHT_GROUPS_MATCHING_ALL = "highlightGroupsMatchingAll";
     public static final String HIGHLIGHT_GROUPS_MATCHING_ANY = "highlightGroupsMatchingAny";
@@ -262,9 +256,7 @@ public class JabRefPreferences {
     // When this should be made possible, the code to inspect is net.sf.jabref.gui.preftabs.LabelPatternPrefTab.storeSettings() -> LabelPattern keypatterns = getLabelPattern(); etc
     public static final String DEFAULT_LABEL_PATTERN = "defaultLabelPattern";
 
-    public static final String SEARCH_ALL_BASES = "searchAllBases";
-    public static final String SHOW_SEARCH_IN_DIALOG = "showSearchInDialog";
-    public static final String FLOAT_SEARCH = "floatSearch";
+    public static final String SEARCH_MODE_FLOAT = "floatSearch";
     public static final String GRAY_OUT_NON_HITS = "grayOutNonHits";
     public static final String CONFIRM_DELETE = "confirmDelete";
     public static final String WARN_BEFORE_OVERWRITING_KEY = "warnBeforeOverwritingKey";
@@ -553,17 +545,12 @@ public class JabRefPreferences {
         defaults.put(MERGEENTRIES_SIZE_Y, 600);
         defaults.put(DEFAULT_SHOW_SOURCE, Boolean.FALSE);
         defaults.put(DEFAULT_AUTO_SORT, Boolean.FALSE);
-        defaults.put(CASE_SENSITIVE_SEARCH, Boolean.FALSE);
-        defaults.put(SEARCH_REQ, Boolean.TRUE);
-        defaults.put(SEARCH_OPT, Boolean.TRUE);
-        defaults.put(SEARCH_GEN, Boolean.TRUE);
-        defaults.put(SEARCH_ALL, Boolean.FALSE);
-        defaults.put(INCREMENT_S, Boolean.FALSE);
-        defaults.put(SEARCH_AUTO_COMPLETE, Boolean.TRUE);
+        defaults.put(SEARCH_CASE_SENSITIVE, Boolean.FALSE);
+        defaults.put(SEARCH_MODE_FILTER, Boolean.TRUE);
 
-        defaults.put(SELECT_S, Boolean.FALSE);
-        defaults.put(REG_EXP_SEARCH, Boolean.TRUE);
-        defaults.put(HIGH_LIGHT_WORDS, Boolean.TRUE);
+        defaults.put(SEARCH_REG_EXP, Boolean.FALSE);
+        defaults.put(SEARCH_PANE_POS_X, 0);
+        defaults.put(SEARCH_PANE_POS_Y, 0);
         defaults.put(EDITOR_EMACS_KEYBINDINGS, Boolean.FALSE);
         defaults.put(EDITOR_EMACS_KEYBINDINGS_REBIND_CA, Boolean.TRUE);
         defaults.put(EDITOR_EMACS_KEYBINDINGS_REBIND_CF, Boolean.TRUE);
@@ -571,7 +558,7 @@ public class JabRefPreferences {
         defaults.put(AUTO_COMPLETE_FIELDS, "author;editor;title;journal;publisher;keywords;crossref");
         defaults.put(AUTO_COMP_FIRST_LAST, Boolean.FALSE); // "Autocomplete names in 'Firstname Lastname' format only"
         defaults.put(AUTO_COMP_LAST_FIRST, Boolean.FALSE); // "Autocomplete names in 'Lastname, Firstname' format only"
-        defaults.put(SHORTEST_TO_COMPLETE, 2);
+        defaults.put(SHORTEST_TO_COMPLETE, 1);
         defaults.put(AUTOCOMPLETE_FIRSTNAME_MODE, JabRefPreferences.AUTOCOMPLETE_FIRSTNAME_MODE_BOTH);
         defaults.put(GROUP_FLOAT_SELECTIONS, Boolean.TRUE);
         defaults.put(GROUP_INTERSECT_SELECTIONS, Boolean.TRUE);
@@ -591,7 +578,6 @@ public class JabRefPreferences {
         defaults.put(HIGHLIGHT_GROUPS_MATCHING_ANY, Boolean.FALSE);
         defaults.put(HIGHLIGHT_GROUPS_MATCHING_ALL, Boolean.FALSE);
         defaults.put(TOOLBAR_VISIBLE, Boolean.TRUE);
-        defaults.put(SEARCH_PANEL_VISIBLE, Boolean.FALSE);
         defaults.put(DEFAULT_ENCODING, "UTF-8");
         defaults.put(GROUPS_VISIBLE_ROWS, 8);
         defaults.put(DEFAULT_OWNER, System.getProperty("user.name"));
@@ -665,9 +651,7 @@ public class JabRefPreferences {
         defaults.put(WARN_BEFORE_OVERWRITING_KEY, Boolean.TRUE);
         defaults.put(CONFIRM_DELETE, Boolean.TRUE);
         defaults.put(GRAY_OUT_NON_HITS, Boolean.TRUE);
-        defaults.put(FLOAT_SEARCH, Boolean.TRUE);
-        defaults.put(SHOW_SEARCH_IN_DIALOG, Boolean.FALSE);
-        defaults.put(SEARCH_ALL_BASES, Boolean.FALSE);
+        defaults.put(SEARCH_MODE_FLOAT, Boolean.FALSE);
         defaults.put(DEFAULT_LABEL_PATTERN, "[authors3][year]");
         defaults.put(PREVIEW_ENABLED, Boolean.TRUE);
         defaults.put(ACTIVE_PREVIEW, 0);