+++ src/main/java/net/sf/jabref/gui/IconTheme.java
@@ -141,6 +141,8 @@ public class IconTheme {
         KEY_BINDINGS("\uf3c6"), /*css: keyboard */
         FIND_DUPLICATES("\uf23e"), /*css: code-equal */
 
+        OPEN_IN_NEW_WINDOW("\uf47b"), /*css: open-in-new */
+
         // STILL MISSING:
         GROUP_REGULAR("\uF4E6", Color.RED);
 
@@ -223,6 +225,11 @@ public class IconTheme {
         public FontBasedIcon createSmallIcon() {
             return new FontBasedIcon(this.iconCode, this.iconColor, IconTheme.SMALL_SIZE);
         }
+
+        public FontBasedIcon createWithNewColor(Color newColor) {
+            return new FontBasedIcon(this.iconCode, newColor, this.size);
+        }
+
     }