+++ src/main/java/net/sf/jabref/logic/autocompleter/ContentAutoCompleters.java
@@ -19,7 +19,7 @@ public class ContentAutoCompleters extends AutoCompleters {
     public ContentAutoCompleters(BibtexDatabase database, MetaData metaData) {
         String[] completeFields = Globals.prefs.getStringArray(JabRefPreferences.AUTO_COMPLETE_FIELDS);
         for (String field : completeFields) {
-            AutoCompleter autoCompleter = AutoCompleterFactory.getFor(field);
+            AutoCompleter<String> autoCompleter = AutoCompleterFactory.getFor(field);
             put(field, autoCompleter);
         }
 
@@ -34,8 +34,8 @@ public class ContentAutoCompleters extends AutoCompleters {
      * values to the autocompleter list:
      */
     public void addContentSelectorValuesToAutoCompleters(MetaData metaData) {
-        for (Map.Entry<String, AutoCompleter> entry : this.autoCompleters.entrySet()) {
-            AutoCompleter ac = entry.getValue();
+        for (Map.Entry<String, AutoCompleter<String>> entry : this.autoCompleters.entrySet()) {
+            AutoCompleter<String> ac = entry.getValue();
             if (metaData.getData(Globals.SELECTOR_META_PREFIX + entry.getKey()) != null) {
                 Vector<String> items = metaData.getData(Globals.SELECTOR_META_PREFIX + entry.getKey());
                 if (items != null) {
@@ -52,7 +52,7 @@ public class ContentAutoCompleters extends AutoCompleters {
      * journal names in the journal abbreviation list to this autocompleter.
      */
     public void addJournalListToAutoCompleter() {
-        AutoCompleter autoCompleter = get("journal");
+        AutoCompleter<String> autoCompleter = get("journal");
         if(autoCompleter != null) {
             for(Abbreviation abbreviation : Abbreviations.journalAbbrev.getAbbreviations()) {
                 autoCompleter.addWordToIndex(abbreviation.getName());